hackmatrix-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.9.0
sqlalchemy>=2.0.23
pyjwt>=2.10.1
bcrypt>=4.1.1
python-multipart>=0.0.6
httpx>=0.25.2
aiofiles>=23.2.1
numpy>=2.0.0
scikit-learn>=1.5.0
email-validator>=2.1.0
feedparser
psycopg2-binary
